pyproject.toml
[project]
name = "whatpeoplewant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exospherehost>=0.0.7b7",
    "pymongo>=4.14.0",
    "python-dotenv>=1.1.1",
]
